Move everything into plugin profile directory
